Modified temporary mapping helper to comply with repo structure
